+++ packages/gov-ui-kit/src/core/components/definitionList/index.ts
@@ -1,0 +1,10 @@
+import { DefinitionListContainer } from './definitionListContainer';
+import { DefinitionListItem } from './definitionListItem';
+
+export const DefinitionList = {
+    Container: DefinitionListContainer,
+    Item: DefinitionListItem,
+};
+
+export * from './definitionListContainer';
+export * from './definitionListItem';