+++ packages/gov-ui-kit/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+-   Implement `DefinitionList` core component
 -   Implement `TransactionDataListItem.Skeleton` module component
 
 ### Changed